requirements.txt
matplotlib
pandas
plotly
streamlit-date-picker